chatbot/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
colored==1.4.2
coverage==5.5
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
eprint==0.0.4
filelock==3.0.12
freezegun==1.1.0
idna==2.10
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.4.1
nltk==3.6.5
numpy==1.20.2
opencv-python==4.5.3.56
peewee==3.14.4
Pillow==8.1.2
pony==0.7.14
psycopg2==2.9.1
psycopg2-binary==2.9.1
pygame==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2021.10.23
requests==2.25.1
scipy==1.6.2
self==2020.12.3
simple-draw==2.7.0
six==1.15.0
soupsieve==2.2.1
termcolor==1.1.0
tqdm==4.62.3
urllib3==1.26.4
virtualenv==20.4.3
vk-api==11.9.1
